app/requirements.txt
# requirements.txt

Flask
Flask-SQLAlchemy
Flask-Migrate==4.1.0
flask-smorest==0.46.1
gunicorn
python-dotenv
